# TINKERPOP-2205 Change connection management to single request per channel