@divijvaidya 2019-05-08
Thank you taking time to review. I know it is quite a big change and might take some time to review.

I would be happy to jump on a call/chat to help you understand the architecture of the changes better. Let me know if that is required.

Regarding the release timeline for this change:
1. Do we have a tentative release timeline for 3.5.0 yet?
2. I will write another create slate version for 3.5 later. Couple of additional things I want to fix as part of the clean slate version is to unify the client API, better DNS resolution for potential hosts in the cluster, additional load balancing strategies, ability for providers to plug in information in the request etc.

I understand that you are not comfortable with releasing such a large change in 3.4.x, but, I am still inclined to release this as part of 3.4.x because in addition to easier configuration, this change also makes the client more robust in handling multiple error scenarios. I would like to see this in the hands of users as soon as possible, else we would have to do additional work to solve problems in existing 3.4.x client.
The only reason I can see to delay this change to 3.5 would be the backward incompatibility with custom channelizes that providers might have written extending from AbstractChannelizer. This can be addressed in the code, if required.

To address your concerns I can do the following:
- Adding additional integration/stress/smoke tests.
- Adding a config parameter to fallback to old code.
- Provide a document on working of the new code prior to merge which can be reviewed by the community for potential bugs/design deficiency.

Let me know how you want to proceed.