@divijvaidya 2019-05-10
> I think that leads me to my problem with merging this to 3.4.x - If someone upgrades from 3.4.1 to 3.4.2 with this change, their configuration options are no longer relevant. They will just be ignored so if the user managed to get their configuration "perfect" under the old model, they now have to change that given this upgrade. Perhaps that's a simple change (maybe??), but it's a change nonetheless.

This is not true. If someone upgrades from 3.4.1 to 3.4.2 with this change, their configuration options would still remain exactly the same. For example, let's say the old configuration was as follows:
```
maxSimultaneousUsage = 32
maxInProcess = 32
maxConnectionPoolSize = 16
(we can ignore the min limits since they are an optimization supporting lazy loading rather than a requirement)
```
This configuration would mean that the user wants to execute 16x32 queries in parallel with each of the 16 connections able to serve 32 queries simultaneously.
In the new code, we decide the max number of connections in the connection pool = [maxConnectionPoolSize + Max(maxSimultaneousUsage, maxInProcess)](https://github.com/apache/tinkerpop/pull/1105/files#diff-0521cde5e923f69a46aad7852b209677R219) which, from a users perspective maintains the same behaviour as the old configuration, i.e. executing 16x32 queries in parallel.

Does this make sense? Do you agree?

> what were you thoughts on what a "switch" would look like?

I will think a bit more about how to do this elegantly but due to the nature of invasive changes, I am afraid the "switch" is going to be a boolean controlling if/else branches across the code.