@spmallette 2019-05-13
> which, from a users perspective maintains the same behaviour as the old configuration, i.e. executing 16x32 queries in parallel.

Thanks for this clarification. I see it now - I agree.

> I am afraid the "switch" is going to be a boolean controlling if/else branches across the code.

hmm - well, unless there aren't many such if/else situations such a change might just make the whole thing messy, but i can't imagine that we'd get lucky that way. Maybe you don't need to try to do all that. How do you feel about keeping this targeted at 3.4.x but not merging for the release of 3.4.2 next week? If we did that, we would have ~45 days of testing, review, etc for 3.4.3 (release would be middle/end of july given our current two month release cycles which I believe would continue).