@spmallette 2019-05-13
> I agree with targeting this change for v3.4.3.

:+1: 

> If any providers have provided a custom channelizer by extending AbstractChannelizer, they will have to modify their channelizer implementation. Is there a way we can inform providers of such implementations?

You should write an entry in the upgrade documentation: 

https://github.com/apache/tinkerpop/blob/master/docs/src/upgrade/release-3.4.x.asciidoc

> Is there anything else left (from an implementation perspective) that you would like me to change? I will address the documentation/examples about the working of the client in a separate PR.

I will be doing some manual tests this week. Since we can't merge this until after we go through our release process, it should be fine to keep pushing changes to this PR (i.e. the documentation updates). I don't really like to break up doc fixes from code changes, if possible. Is there a reason you would like to see it merge ahead of those changes? since we're waiting for 3.4.3 now there is a bit of a delay to merge anyway right?