@divijvaidya 2019-05-13
No, there is no reason to merge this before the 3.4.2 release. We can merge it after that. Meanwhile, I will continue adding changes to this branch.