@divijvaidya 2019-06-06
1. I have added the tests as discussed in the JIRA associated with this PR.
2. There is an issue with Netty being tracked in https://github.com/netty/netty/pull/9226 which might cause some weird behaviour while closing the client. I am on it right now and will fix it separately from this PR.